优化

+++ pay-center-alipay/src/main/java/code/ponfee/alipay/AlipayService.java
@@ -456,7 +456,7 @@ public class AlipayService extends PayServiceAdapter {
         map.put("body", params.get(PARAM_GOODS_BODY));
 
         Date expireTime = Dates.toDate(params.get(PARAM_EXPIRE_TIME), DEFAULT_DATE_FORMAT);
-        int seconds = (int) Dates.clockdiff(new Date(), expireTime) / 60;
+        int seconds = Dates.clockdiff(new Date(), expireTime) / 60;
         map.put("expireTime", seconds + "m");
 
         // webPay参数
